refactor: rename domain errors and settle linter configuration
- InvalidTransition / InvalidAmount получили суффикс Error по N818
- RUF001-003 отключены: правила про «неоднозначные» кириллические
  символы рассчитаны на базу, где кириллица означает подмену символов,
  а здесь весь текст русский
- webhook_backoff приведён к float явно: mypy strict видел Any
  в возведении в степень

Весь набор после рефакторинга: 13 passed, ruff и mypy чисто.

+++ app/application/backoff.py
@@ -17,4 +17,4 @@ WEBHOOK_BASE_DELAY = 1.0
 
 def webhook_backoff(attempt: int) -> float:
     """Задержка перед попыткой `attempt + 1` доставки webhook, секунды (RFC §6.2)."""
-    return WEBHOOK_BASE_DELAY * 2 ** (attempt - 1)
+    return float(WEBHOOK_BASE_DELAY * 2 ** (attempt - 1))
